Convert object to D3 data

Generic method to transform R objects into D3 friendly data.

as_d3_data(x, ...)

# S3 method for default
as_d3_data(x, ...)

Arguments

x
data

...
Additional arguments for generic methods

Details

The value returned from as_d3_data() should be one of:

An R data frame. In this case the HTMLWidgets.dataframeToD3() JavaScript function will be called on the client to transform the data into D3 friendly (row-oriented) data; or

A JSON object created using jsonlite::toJSON; or

Any other R object which can be coverted to JSON using jsonlite::toJSON.