Save a D3 visualization as HTML

Save a D3 visualization to an HTML file (e.g. for sharing with others).

save_d3_html(
  d3,
  file,
  selfcontained = TRUE,
  libdir = NULL,
  background = "white",
  title = "D3 Visualization",
  knitrOptions = list()
)

Arguments

d3
D3 visualization to save

file
File to save HTML into

selfcontained
Whether to save the HTML as a single self-contained file (with external resources base64 encoded) or a file with external resources placed in an adjacent directory.

libdir
Directory to copy HTML dependencies into (defaults to filename_files).

background
Text string giving the html background color of the widget. Defaults to white.

title
Text to use as the title of the generated page.

knitrOptions
A list of knitr chunk options.

Details

Using selfcontained set to TRUE requires pandoc to be installed.

See also

Examples

library(r2d3)

viz <- r2d3(
  data = c(0.3, 0.6, 0.8, 0.95, 0.40, 0.20),
  script = system.file("examples/barchart.js", package = "r2d3")
)

save_d3_html(
  viz,
  file = tempfile(fileext = ".html"),
  selfcontained = FALSE
)